+++ lib/plugin/store/balances.dart
@@ -43,13 +43,6 @@ abstract class BalancesStoreBase with Store {
       });
     }
 
-    data.removeWhere((e) => e.symbol!.contains('-') && e.amount == '0');
-    data.sort((a, b) => a.symbol!.contains('-')
-        ? 1
-        : b.symbol!.contains('-')
-            ? -1
-            : a.symbol!.compareTo(b.symbol!));
-
     tokens = data;
     isTokensFromCache = isFromCache;
   }
@@ -82,6 +75,7 @@ class TokenBalanceData {
     this.symbol,
     this.type = 'Token',
     this.currencyId,
+    this.src,
     this.minBalance,
     this.fullName,
     this.decimals,
@@ -98,6 +92,7 @@ class TokenBalanceData {
   final String? symbol;
   final String type;
   final Map? currencyId;
+  final Map? src;
   final String? minBalance;
   final String? fullName;
   final int? decimals;