+++ lib/service/webViewRunner.dart
@@ -28,6 +28,7 @@ class WebViewRunner {
     Keyring keyringStorage,
     Function? onLaunched, {
     String? jsCode,
+    Function? socketDisconnectedAction,
   }) async {
     /// reset state before webView launch or reload
     _msgHandlers = {};
@@ -60,6 +61,10 @@ class WebViewRunner {
             } else {
               jsCodeStarted = 0;
             }
+          }
+          if (message.message.contains("WebSocket is not connected") &&
+              socketDisconnectedAction != null) {
+            socketDisconnectedAction();
           }
           if (message.messageLevel != ConsoleMessageLevel.LOG) return;